requirements.txt
cycler==0.11.0
fonttools==4.33.3
kiwisolver==1.4.2
matplotlib==3.5.2
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pyaml==21.10.1
pyparsing==3.0.9
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
